refactor: extract abstract class AbstractSecurityContextInterceptor
